Robots do not forget where they are if the camera doesn't detect them.
Added hardcoded (hardbuilt?) tags so that the only valid ones are the
first three of all valid tags. So now if the camera doesn't detect them
it still adds blank tags inside found_tags to maintain 3.
Also if a robot does not receive a valid tag its position message is not sent.